backend/requirements.txt
fastapi
httpx
pydantic
python-dotenv
uvicorn
groq
